+++ packages/framer-motion/src/easing/circ.ts
@@ -4,4 +4,4 @@ import { EasingFunction } from "./types"
 
 export const circIn: EasingFunction = (p) => 1 - Math.sin(Math.acos(p))
 export const circOut = reverseEasing(circIn)
-export const circInOut = mirrorEasing(circOut)
+export const circInOut = mirrorEasing(circIn)